refactor(e2e): consolidate seed-uuid constants into fixtures/seed-uuids.ts
Commit 47e1773 unintentionally created `apps/web/e2e/fixtures/seed-uuids.ts`
without noticing the existing `apps/web/e2e/helpers/seed-ids.ts` (Phase 15.1
artifact). Same constants, same values — two parallel sources of truth would
drift on the next seed change.

Merge: keep `fixtures/seed-uuids.ts` (the newer, better-commented module
with full Phase 17.1 context), backfill all 28 exports from the deleted
helper, migrate the 5 importers (`admin-dsgvo-{retention,export-job,vvz,
consents,dsfa}.spec.ts`), delete `helpers/seed-ids.ts`.

Type-check passes (exit 0).
