+++ apps/web/e2e/admin-dsgvo-consents.spec.ts
@@ -27,7 +27,7 @@
 import { test, expect } from '@playwright/test';
 import { loginAsAdmin } from './helpers/login';
 import { seedConsent } from './helpers/dsgvo';
-import { SEED_PERSON_STUDENT_1_UUID } from './helpers/seed-ids';
+import { SEED_PERSON_STUDENT_1_UUID } from './fixtures/seed-uuids';
 
 test.describe.configure({ mode: 'serial' });
