+++ apps/web/e2e/admin-dsgvo-retention.spec.ts
@@ -25,7 +25,7 @@
 import { test, expect } from '@playwright/test';
 import { loginAsAdmin } from './helpers/login';
 import { seedRetentionPolicy, cleanupAll } from './helpers/dsgvo';
-import { SEED_SCHOOL_UUID } from './helpers/seed-ids';
+import { SEED_SCHOOL_UUID } from './fixtures/seed-uuids';
 
 test.describe.configure({ mode: 'serial' });
