+++ apps/web/e2e/admin-dsgvo-vvz.spec.ts
@@ -10,7 +10,7 @@
 import { test, expect } from '@playwright/test';
 import { loginAsAdmin } from './helpers/login';
 import { seedVvzEntry, cleanupAll } from './helpers/dsgvo';
-import { SEED_SCHOOL_UUID } from './helpers/seed-ids';
+import { SEED_SCHOOL_UUID } from './fixtures/seed-uuids';
 
 test.describe.configure({ mode: 'serial' });
